Split crypto.hpp into 3 different files
