+++ llarp/crypto/constants.cpp
@@ -1,0 +1,1 @@
+#include <crypto/constants.hpp>
