+++ llarp/dht/bucket.hpp
@@ -4,6 +4,7 @@
 #include <crypto/crypto.hpp>
 #include <dht/kademlia.hpp>
 #include <dht/key.hpp>
+
 #include <map>
 #include <set>
 #include <vector>
