+++ llarp/link/encoder.hpp
@@ -1,13 +1,13 @@
 #ifndef LLARP_LINK_ENCODER_HPP
 #define LLARP_LINK_ENCODER_HPP
 
-#include <router_contact.hpp>
-
-#include <util/bencode.h>
-#include <util/buffer.h>
+struct llarp_buffer_t;
 
 namespace llarp
 {
+  struct KeyExchangeNonce;
+  class RouterContact;
+
   /// encode Link Introduce Message onto a buffer
   /// if router is nullptr then the LIM's r member is omitted.
   bool
