+++ llarp/link/iwp_internal.hpp
@@ -1,5 +1,8 @@
 #ifndef LLARP_LINK_IWP_INTERNAL_HPP
 #define LLARP_LINK_IWP_INTERNAL_HPP
+
+#include <crypto/crypto.hpp>
+#include <crypto/types.hpp>
 #include <link/server.hpp>
 #include <link/session.hpp>
 #include <link_layer.hpp>
@@ -9,6 +12,7 @@
 
 namespace llarp
 {
+  struct Crypto;
   namespace iwp
   {
     struct LinkLayer;
