+++ llarp/messages/exit.hpp
@@ -1,7 +1,7 @@
 #ifndef LLARP_MESSAGES_EXIT_HPP
 #define LLARP_MESSAGES_EXIT_HPP
 
-#include <crypto/crypto.hpp>
+#include <crypto/types.hpp>
 #include <exit/policy.hpp>
 #include <routing/message.hpp>
 
@@ -9,6 +9,7 @@
 
 namespace llarp
 {
+  struct Crypto;
   namespace routing
   {
     struct ObtainExitMessage final : public IMessage
