+++ llarp/messages/link_intro.hpp
@@ -1,6 +1,7 @@
 #ifndef LLARP_MESSAGES_LINK_INTRO_HPP
 #define LLARP_MESSAGES_LINK_INTRO_HPP
 
+#include <crypto/types.hpp>
 #include <link_message.hpp>
 #include <router_contact.hpp>
