+++ llarp/messages/path_transfer.hpp
@@ -1,7 +1,7 @@
 #ifndef LLARP_MESSAGES_PATH_TRANSFER_HPP
 #define LLARP_MESSAGES_PATH_TRANSFER_HPP
 
-#include <crypto/crypto.hpp>
+#include <crypto/types.hpp>
 #include <encrypted.hpp>
 #include <routing/message.hpp>
 #include <service/protocol.hpp>
