+++ llarp/messages/relay.hpp
@@ -1,7 +1,7 @@
 #ifndef LLARP_MESSAGES_RELAY_HPP
 #define LLARP_MESSAGES_RELAY_HPP
 
-#include <crypto/crypto.hpp>
+#include <crypto/types.hpp>
 #include <encrypted.hpp>
 #include <link_message.hpp>
 #include <path/path_types.hpp>
