+++ llarp/service/handler.hpp
@@ -1,7 +1,7 @@
 #ifndef LLARP_SERVICE_HANDLER_HPP
 #define LLARP_SERVICE_HANDLER_HPP
 
-#include <crypto/crypto.hpp>
+#include <crypto/types.hpp>
 #include <path/path_types.hpp>
 #include <service/IntroSet.hpp>
 #include <util/aligned.hpp>
