+++ test/routing/test_llarp_routing_obtainexitmessage.cpp
@@ -1,5 +1,6 @@
 #include <gtest/gtest.h>
 
+#include <crypto/crypto.hpp>
 #include <messages/exit.hpp>
 
 using ObtainExitMessage = llarp::routing::ObtainExitMessage;
